+++ modules/postgres/app/edu/berkeley/ground/postgres/controllers/GraphController.java
@@ -58,7 +58,7 @@ public class GraphController extends Controller {
       () -> {
         try {
           return this.cache.getOrElse(
-            "graphs",
+            "graphs." + sourceKey,
             () -> Json.toJson(this.postgresGraphDao.retrieveFromDatabase(sourceKey)),
             Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
         } catch (Exception e) {
@@ -74,7 +74,7 @@ public class GraphController extends Controller {
     return CompletableFuture.supplyAsync(
       () -> {
         try {
-          return this.cache.getOrElse("graph_versions",
+          return this.cache.getOrElse("graph_versions." + id,
             () -> Json.toJson(this.postgresGraphVersionDao.retrieveFromDatabase(id)),
             Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
         } catch (Exception e) {
@@ -121,6 +121,40 @@ public class GraphController extends Controller {
           throw new CompletionException(e);
         }
         return Json.toJson(graphVersion);
+      },
+      PostgresUtils.getDbSourceHttpContext(actorSystem))
+             .thenApply(Results::ok)
+             .exceptionally(e -> GroundUtils.handleException(e, request()));
+  }
+
+  public final CompletionStage<Result> getLatest(String sourceKey) {
+    return CompletableFuture.supplyAsync(
+      () -> {
+        try {
+          return this.cache.getOrElse(
+            "graph_leaves." + sourceKey,
+            () -> Json.toJson(this.postgresGraphDao.getLeaves(sourceKey)),
+            Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
+        } catch (Exception e) {
+          throw new CompletionException(e);
+        }
+      },
+      PostgresUtils.getDbSourceHttpContext(actorSystem))
+             .thenApply(Results::ok)
+             .exceptionally(e -> GroundUtils.handleException(e, request()));
+  }
+
+  public final CompletionStage<Result> getHistory(String sourceKey) {
+    return CompletableFuture.supplyAsync(
+      () -> {
+        try {
+          return this.cache.getOrElse(
+            "graph_history." + sourceKey,
+            () -> Json.toJson(this.postgresGraphDao.getHistory(sourceKey)),
+            Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
+        } catch (Exception e) {
+          throw new CompletionException(e);
+        }
       },
       PostgresUtils.getDbSourceHttpContext(actorSystem))
              .thenApply(Results::ok)