+++ modules/postgres/app/edu/berkeley/ground/postgres/controllers/NodeController.java
@@ -46,7 +46,7 @@ public class NodeController extends Controller {
       () -> {
         try {
           return this.cache.getOrElse(
-            "nodes",
+            "nodes." + sourceKey,
             () -> Json.toJson(this.postgresNodeDao.retrieveFromDatabase(sourceKey)),
             Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
         } catch (Exception e) {
@@ -81,7 +81,7 @@ public class NodeController extends Controller {
       () -> {
         try {
           return this.cache.getOrElse(
-            "node_versions",
+            "node_versions." + id,
             () -> Json.toJson(this.postgresNodeVersionDao.retrieveFromDatabase(id)),
             Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
         } catch (Exception e) {
@@ -115,4 +115,56 @@ public class NodeController extends Controller {
              .thenApply(Results::created)
              .exceptionally(e -> GroundUtils.handleException(e, request()));
   }
+
+  public final CompletionStage<Result> getLatest(String sourceKey) {
+    return CompletableFuture.supplyAsync(
+      () -> {
+        try {
+          return this.cache.getOrElse(
+            "node_leaves." + sourceKey,
+            () -> Json.toJson(this.postgresNodeDao.getLeaves(sourceKey)),
+            Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
+        } catch (Exception e) {
+          throw new CompletionException(e);
+        }
+      },
+      PostgresUtils.getDbSourceHttpContext(actorSystem))
+             .thenApply(Results::ok)
+             .exceptionally(e -> GroundUtils.handleException(e, request()));
+  }
+
+  public final CompletionStage<Result> getHistory(String sourceKey) {
+    return CompletableFuture.supplyAsync(
+      () -> {
+        try {
+          return this.cache.getOrElse(
+            "node_history." + sourceKey,
+            () -> Json.toJson(this.postgresNodeDao.getHistory(sourceKey)),
+            Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
+        } catch (Exception e) {
+          throw new CompletionException(e);
+        }
+      },
+      PostgresUtils.getDbSourceHttpContext(actorSystem))
+             .thenApply(Results::ok)
+             .exceptionally(e -> GroundUtils.handleException(e, request()));
+  }
+
+  public final CompletionStage<Result> getAdjacentLineage(Long id) {
+    return CompletableFuture.supplyAsync(
+      () -> {
+        try {
+          return this.cache.getOrElse(
+            "node_version_adj_lineage." + id,
+            () -> Json.toJson(this.postgresNodeVersionDao.retrieveAdjacentLineageEdgeVersion(id)),
+            Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
+        } catch (Exception e) {
+          throw new CompletionException(e);
+        }
+      },
+      PostgresUtils.getDbSourceHttpContext(actorSystem))
+             .thenApply(Results::ok)
+             .exceptionally(e -> GroundUtils.handleException(e, request()));
+  }
+
 }