+++ modules/postgres/app/edu/berkeley/ground/postgres/controllers/StructureController.java
@@ -57,7 +57,7 @@ public class StructureController extends Controller {
       () -> {
         try {
           return this.cache.getOrElse(
-            "structures",
+            "structures." + sourceKey,
             () -> Json.toJson(this.postgresStructureDao.retrieveFromDatabase(sourceKey)),
             Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
         } catch (Exception e) {
@@ -73,7 +73,7 @@ public class StructureController extends Controller {
       () -> {
         try {
           return this.cache.getOrElse(
-            "structure_versions",
+            "structure_versions." + id,
             () -> Json.toJson(this.postgresStructureVersionDao.retrieveFromDatabase(id)),
             Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
         } catch (Exception e) {
@@ -124,6 +124,40 @@ public class StructureController extends Controller {
       },
       PostgresUtils.getDbSourceHttpContext(this.actorSystem))
              .thenApply(Results::created)
+             .exceptionally(e -> GroundUtils.handleException(e, request()));
+  }
+
+  public final CompletionStage<Result> getLatest(String sourceKey) {
+    return CompletableFuture.supplyAsync(
+      () -> {
+        try {
+          return this.cache.getOrElse(
+            "structure_leaves." + sourceKey,
+            () -> Json.toJson(this.postgresStructureDao.getLeaves(sourceKey)),
+            Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
+        } catch (Exception e) {
+          throw new CompletionException(e);
+        }
+      },
+      PostgresUtils.getDbSourceHttpContext(actorSystem))
+             .thenApply(Results::ok)
+             .exceptionally(e -> GroundUtils.handleException(e, request()));
+  }
+
+  public final CompletionStage<Result> getHistory(String sourceKey) {
+    return CompletableFuture.supplyAsync(
+      () -> {
+        try {
+          return this.cache.getOrElse(
+            "structure_history." + sourceKey,
+            () -> Json.toJson(this.postgresStructureDao.getHistory(sourceKey)),
+            Integer.parseInt(System.getProperty("ground.cache.expire.secs")));
+        } catch (Exception e) {
+          throw new CompletionException(e);
+        }
+      },
+      PostgresUtils.getDbSourceHttpContext(actorSystem))
+             .thenApply(Results::ok)
              .exceptionally(e -> GroundUtils.handleException(e, request()));
   }
 }