+++ resources/ci/init.sh
@@ -5,8 +5,12 @@ sudo service postgresql start
 sudo service postgresql status
 sudo service postgresql restart
 
-# create Postgres db and user
-sudo su -c "createdb test" -s /bin/sh postgres
-sudo su -c "createuser test -d -s" -s /bin/sh postgres
+echo "restarted postgres"
 
-cd resources/scripts/postgres && python2.7 postgres_setup.py test test && cd ../../..
+# create Postgres db and user
+sudo su -c "createdb test" -s /bin/bash postgres
+sudo su -c "createuser test -d -s" -s /bin/bash postgres
+
+echo "created dbs"
+
+cd modules/postgres/dist/db && python2.7 postgres_setup.py test test && cd ../../../..