+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 # Changelog
 
-## [0.1.1] - 2017-06-22
+##[0.1.2](https://github.com/ground-context/ground/releases/tag/v0.1.2)
+* Add API to retrieve the latest version(s) of an `Item`.
+* Add API to retrieve the history of an `Item`. 
+* Add API that retrieves lineage edge versions adjacent to a particular node version.
 
+## [0.1.1](https://github.com/ground-context/ground/releases/tag/v0.1.1)
 ### Fixed
 
 * Fixed bug that prevented the creation of lineage graphs.
 * Fixed `null` serialization issue for some strings.
 * Removed required empty data structures (i.e., the user no longer has to provide an empty `tags` map if no tags are being added).
 
-## [0.1.0] - 2017-05-22
+## [0.1.0](https://github.com/ground-context/ground/releases/tag/v0.1.0)
 
 * Initial Release of Ground.